\name{data1}
\alias{data1}
\docType{data}

\title{
Example data for survival.QC1
}

\description{
This is a data for \code{survival.QC1}. It contains survival time and 
censoring indicator.
}

\usage{data(simul1)}

\format{
  A data frame with 50 by 2 matrix.
  \describe{
    \item{\code{data}}{
    A 50 by 2 matrix.  The first column denotes survival time and 
    the second column denotes censoring indicator taking 1 for the 
    observed survival time and 0 for the censored survival time.
    }
  }
}

\references{

}

\examples{

}

\keyword{datasets}
