\name{data2}
\alias{data2}
\docType{data}

\title{
Example data for survival.QC2
}

\description{
This is a data for \code{survival.QC2}. It contains survival time and 
accrual time.
}

\usage{data(simul2)}

\format{
  A data frame with 50 by 2 matrix.
  \describe{
    \item{\code{data}}{
    A 50 by 2 matrix.  The first column denotes survival time and 
    the second column denotes accrual time.
    }
  }
}

\references{

}

\examples{

}

\keyword{datasets}
