\name{QC4CT-package}
\alias{QC4CT-package}
\alias{QC4CT}
\docType{package}
\title{
\packageTitle{QC4CT}
}
\description{
\packageDescription{QC4CT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{QC4CT}
\packageIndices{QC4CT}

There are three main functions. Function \code{accrual.QC} provides test statistics and p-value of Kolmogorov-Smirnov test for the identification of accrual distribution. Functions \code{survival.QC1} and \code{survival.QC2} provide weighted Kaplan-Meier statistics of Pepe and Fleming (1989) performing distance test for the identification of survival distribution. \code{survival.QC1} works when the observed survival time and censoring indicator are given; and \code{survival.QC2} works when the observed survival time and accrual time are given. 
}

\author{
\packageAuthor{QC4CT}

Maintainer: \packageMaintainer{QC4CT}
}
\references{
Pepe, M. S. and Fleming, T. R. (1989). Weighted Kaplan-Meier statistics: a class of distance tests for censored survival data. Biometrics, 497-507.
}
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
